server/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
matplotlib==3.9.2 
github==1.2.7
requests==2.32.3
networkx==3.4.2
GitPython == 3.1.43